requirements.txt
absl-py==1.4.0
addict==2.4.0
ansi2html==1.8.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
asttokens==2.4.0
attrs==23.1.0
backcall==0.2.0
Brotli
cachetools==5.3.1
certifi==2023.7.22
charset-normalizer
click==8.1.7
comm==0.1.4
ConfigArgParse==1.7
contourpy==1.1.0
cycler==0.11.0
dash==2.13.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
docker-pycreds==0.4.0
easydict==1.10
executing==1.2.0
fastjsonschema==2.18.0
filelock==3.12.4
Flask==2.2.5
fonttools==4.42.1
fsspec==2023.9.2
gitdb==4.0.11
GitPython==3.1.41
gmpy2
google-auth==2.22.0
google-auth-oauthlib==1.0.0
grpcio==1.57.0
huggingface-hub==0.17.3
idna
imageio==2.26.0
imageio-ffmpeg==0.4.8
importlib-metadata==6.8.0
importlib-resources==6.0.1
install==1.3.5
ipython==8.12.2
ipywidgets==8.1.0
itsdangerous==2.1.2
jedi==0.19.0
Jinja2
joblib==1.3.2
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jupyter_core==5.3.1
jupyterlab-widgets==3.0.8
kiwisolver==1.4.5
kornia==0.6.10
lpips==0.1.4
Markdown==3.4.4
markdown-it-py==3.0.0
MarkupSafe
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdurl==0.1.2
mkl-fft==1.3.6
mkl-random==1.2.2
mkl-service==2.4.0
mpmath
nbformat==5.7.0
nest-asyncio==1.5.7
networkx
numpy==1.24.3
nvidia-ml-py==12.535.133
nvitop==1.3.1
oauthlib==3.2.2
olefile
omegaconf==2.3.0
open3d==0.17.0
opencv-python==4.7.0.72
packaging==23.1
pandas==2.0.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
plotly==5.16.1
plyfile==1.0.1
prompt-toolkit==3.0.39
protobuf==4.24.2
psutil==5.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
Pygments==2.16.1
pyparsing==3.0.9
pyquaternion==0.9.9
PySocks
python-dateutil==2.8.2
pytorch-msssim==0.2.1
pytz==2023.3.post1
PyWavelets==1.4.1
PyYAML
referencing==0.30.2
requests
requests-oauthlib==1.3.1
retrying==1.3.4
rich==13.6.0
rpds-py==0.10.2
rsa==4.9
safetensors==0.3.3
scikit-image==0.19.3
scikit-learn==1.3.0
scipy==1.11.0
sentry-sdk==1.40.1
setproctitle==1.3.3
six
smmap==5.0.1
stack-data==0.6.2
sympy
tenacity==8.2.3
tensorboard==2.14.0
tensorboard-data-server==0.7.1
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2023.7.10
timm==0.9.7
tqdm==4.65.0
traitlets==5.9.0
triton==2.1.0
typing_extensions
tzdata==2023.3
urllib3==1.26.16
wandb==0.16.2
wcwidth==0.2.6
Werkzeug==2.2.3
widgetsnbextension==4.0.8
zipp==3.16.2